+++ src/main/java/com/bvk/partner/tokopedia/seller/object/UpdateShopStatus.java
@@ -2,7 +2,7 @@ package com.bvk.partner.tokopedia.seller.object;
 
 public class UpdateShopStatus {
 
-	public Integer shop_id;
+	public Long shop_id;
     public String action;
     public String start_date;
     public String end_date;
